Automatically check for app updates on startup
