+++ src/aohara/tinkertime/workflows/ModWorkflowBuilder.java
@@ -18,6 +18,9 @@ import aohara.tinkertime.crawlers.VersionInfo;
 import aohara.tinkertime.models.FileUpdateListener;
 import aohara.tinkertime.models.Mod;
 import aohara.tinkertime.models.ModStructure;
+import aohara.tinkertime.workflows.contexts.DirectDownloaderContext;
+import aohara.tinkertime.workflows.contexts.DownloaderContext;
+import aohara.tinkertime.workflows.contexts.ModDownloaderContext;
 import aohara.tinkertime.workflows.tasks.CacheCrawlerPageTask;
 import aohara.tinkertime.workflows.tasks.CheckForUpdateTask;
 import aohara.tinkertime.workflows.tasks.CrawlerDownloadTask;
